override Dataset __setattr__ function
With this commit, it is possible to do Dataset.name = value instead of
the usual Dataset['name'] = (coords, value).

There may be a good reason that this is currently blocked by xarray,
but none immediately came to mind so I thought this would be worth
trying.

Another alternative to this is to do Dataset.name.data = value if your
'name' DataArray was called 'data'. This is already possible with the
default xarray Dataset object.